+++ dev/ViewModel/TheHeroViMo.cs
@@ -166,7 +166,7 @@ namespace FateExplorer.ViewModel
         public string PlaceOfBirth { get => characterM?.PlaceOfBirth ?? ""; }
         public string DateOfBirth { get => characterM?.DateOfBirth ?? ""; }
 
-        public double CarriedWeight { get => characterM?.CarriedWeight ?? 0; }
+        public decimal CarriedWeight { get => characterM?.CarriedWeight ?? 0; }
 
 
         #region POSESSIONS ------
@@ -212,9 +212,9 @@ namespace FateExplorer.ViewModel
             return Result;
         }
 
-        public double WhatCanCarry { get => characterM?.WhatCanCarry(AbilityEffValues[AbilityM.STR]) ?? 0; }
+        public decimal WhatCanCarry { get => characterM?.WhatCanCarry(AbilityEffValues[AbilityM.STR]) ?? 0; }
 
-        public double WhatCanLift { get => characterM?.WhatCanLift(AbilityEffValues[AbilityM.STR]) ?? 0; }
+        public decimal WhatCanLift { get => characterM?.WhatCanLift(AbilityEffValues[AbilityM.STR]) ?? 0; }
 
         #endregion
